core: sync refinements (#531)
* update files and folder docs

* further document sync

* non functional refactor to make sync human readable

* reorder sync to have the most useful things first

* fixed a test to be more reasonable

* fixed db comparison

* write a failing test

* fix tests that hardcoded which db they want to use

* we do want these container names

* deleteing a path that does not exist is not an error

* cleanup artifacts of old code structure

+++ core/src/storage/db_provider.rs
@@ -147,7 +147,11 @@ impl Backend for FileBackend {
         let path_str = Self::key_path(db, namespace, key);
         let path = Path::new(&path_str);
         trace!("delete\t{}", &path_str);
-        remove_file(path)
+        if path.exists() {
+            remove_file(path)
+        } else {
+            Ok(())
+        }
     }
 
     fn dump<N, V>(db: &Self::Db, namespace: N) -> Result<Vec<V>, Self::Error>
